requirements.txt
pillow
numpy
joblib
cupy-cuda111
